+++ onboard-jmuniz/app/(home)/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from "expo-router";
+
+export default function HomeLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="list" options={{ title: "Taqtile Users" }} />
+    </Stack>
+  );
+}
