--- onboard-jmuniz/app/home/_layout.tsx
@@ -1,9 +1,0 @@
-import { Stack } from "expo-router";
-
-export function HomeLayout() {
-  return (
-    <Stack>
-      <Stack.Screen name="index" options={{ title: "Title" }} />
-    </Stack>
-  );
-}
