--- onboard-jmuniz/app/home/title.tsx
@@ -1,5 +1,0 @@
-import { Text } from "react-native";
-
-export default function Header() {
-  return <Text>Taqtile Users</Text>;
-}
